chore: configure path alias and update import statements to use '@'
