+++ src/entities/comment/ui/CommentLikeButton.tsx
@@ -1,5 +1,5 @@
 import { ThumbsUp } from "lucide-react"
-import { Button } from "../../../shared/ui"
+import { Button } from "@/shared/ui"
 import useCommentButton from "../model/useCommentButton"
 
 interface CommentLikeButtonProps {
