+++ src/entities/comment/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./CommentLikeButton"
+export * from "./CommentEditButton"
+export * from "./CommentDeleteButton"
