+++ src/features/comment/ui/CommentAddDialog.tsx
@@ -1,4 +1,4 @@
-import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from "../../../shared/ui"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from "@/shared/ui"
 import useComments from "../model/useComments"
 
 /* 댓글 추가 대화상자 */
