+++ src/features/post/ui/PostDetailDialog.tsx
@@ -1,7 +1,8 @@
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
-import { HightLightText } from "../../../shared/ui/HighLightText"
+import { CommentSection } from "@/widgets/ui/CommentSection"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/shared/ui"
+import { HightLightText } from "@/shared/ui/HighLightText"
 import usePosts from "../model/usePosts"
-import { CommentSection } from "../../../widgets/ui/CommentSection"
+
 
 /* 게시물 상세 보기 대화상자 */
 export const PostDetailDialog = () => {
