+++ src/features/post/ui/PostEditDialog.tsx
@@ -1,4 +1,4 @@
-import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "../../../shared/ui"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "@/shared/ui"
 import usePosts from "../model/usePosts"
 
 /* 게시물 수정 대화상자 */
