+++ vite.config.ts
@@ -1,9 +1,18 @@
 import { defineConfig } from "vite"
 import react from "@vitejs/plugin-react"
+import path, { dirname } from "path"
+import { fileURLToPath } from "url"
+
+const __dirname = dirname(fileURLToPath(import.meta.url))
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "src"),
+    },
+  },
   server: {
     proxy: {
       "/api": {
