+++ vitest.config.ts
@@ -1,9 +1,19 @@
 /// <reference types="vitest" />
 import { defineConfig } from "vitest/config"
 import react from "@vitejs/plugin-react"
+import path, { dirname } from "path"
+import { fileURLToPath } from "url"
+
+const __dirname = dirname(fileURLToPath(import.meta.url))
+
 
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "src"),
+    },
+  },
   test: {
     globals: true,
     environment: "jsdom",
